+++ tests/mocks/now-playlist-track.mocks.ts
@@ -1,0 +1,219 @@
+export let VideoMock = {
+  "kind": "youtube#video",
+  "etag": "\"m2yskBQFythfE4irbTIeOgYYfBU/_lsPlA3UVBxJ-nFZoh_VecCl4jo\"",
+  "id": "CpzTC5fOeZU",
+  "snippet": {
+    "publishedAt": "2015-12-03T19:33:15.000Z",
+    "channelId": "UCsbWPmIwbmnvM-DykpEfwzA",
+    "title": "Ozric Tentacles   Technicians of the Sacred 2015, Full Album",
+    "description": "Purchase the album here\nhttp://www.amazon.com/Technicians-The...\nhttps://www.burningshed.com/store/mad...\n\n00:00 The High Pass\n08:23 Butterfly Garden\n13:27 Far Memory \n20:37 Changa Masala\n26:41 Zingbong\n35:07 Switchback\n\n45:18 Epiphlioy\n57:07 The Unusual Village\n1:03:27 Smiling Potion\n1:10:39 Rubbing Shoulders with the Absolute\n1:19:15 Zenlike Creature",
+    "thumbnails": {
+      "default": {
+        "url": "https://i.ytimg.com/vi/CpzTC5fOeZU/default.jpg",
+        "width": 120,
+        "height": 90
+      },
+      "medium": {
+        "url": "https://i.ytimg.com/vi/CpzTC5fOeZU/mqdefault.jpg",
+        "width": 320,
+        "height": 180
+      },
+      "high": {
+        "url": "https://i.ytimg.com/vi/CpzTC5fOeZU/hqdefault.jpg",
+        "width": 480,
+        "height": 360
+      },
+      "standard": {
+        "url": "https://i.ytimg.com/vi/CpzTC5fOeZU/sddefault.jpg",
+        "width": 640,
+        "height": 480
+      },
+      "maxres": {
+        "url": "https://i.ytimg.com/vi/CpzTC5fOeZU/maxresdefault.jpg",
+        "width": 1280,
+        "height": 720
+      }
+    },
+    "channelTitle": "Ozric Tentacles Official",
+    "tags": [
+      "Technicians Of The Sacred",
+      "Ozric Tentacles (Musical Group)",
+      "Psychedelic Rock (Musical Genre)",
+      "Album",
+      "Song",
+      "New",
+      "Full"
+    ],
+    "categoryId": "22",
+    "liveBroadcastContent": "none",
+    "localized": {
+      "title": "Ozric Tentacles   Technicians of the Sacred 2015, Full Album",
+      "description": "Purchase the album here\nhttp://www.amazon.com/Technicians-The...\nhttps://www.burningshed.com/store/mad...\n\n00:00 The High Pass\n08:23 Butterfly Garden\n13:27 Far Memory \n20:37 Changa Masala\n26:41 Zingbong\n35:07 Switchback\n\n45:18 Epiphlioy\n57:07 The Unusual Village\n1:03:27 Smiling Potion\n1:10:39 Rubbing Shoulders with the Absolute\n1:19:15 Zenlike Creature"
+    }
+  },
+  "contentDetails": {
+    "duration": "PT1H29M21S",
+    "dimension": "2d",
+    "definition": "hd",
+    "caption": "false",
+    "licensedContent": false,
+    "projection": "rectangular"
+  },
+  "statistics": {
+    "viewCount": "8756",
+    "likeCount": "100",
+    "dislikeCount": "0",
+    "favoriteCount": "0",
+    "commentCount": "12"
+  }
+}
+export let VideoMockWithSpecialChars = {
+  "kind": "youtube#video",
+  "etag": "\"m2yskBQFythfE4irbTIeOgYYfBU/NYCHlBzdQXNN3V9f5r02OgaM2m8\"",
+  "id": "j3saYjV6UtA",
+  "snippet": {
+    "publishedAt": "2016-08-14T14:32:38.000Z",
+    "channelId": "UCc8S4NpVjngtbD7jrlWlsIw",
+    "title": "Porcupine Tree - Lightbulb Sun (Full Album Lyrics)",
+    "description": "Here are the lyrics to the entire album of Lightbulb Sun by Porcupine Tree.\n\nLyrics taken from http://www.darklyrics.com/p/porcupinetree.html\n\nTracklist:\n1. Lightbulb Sun [00:00]\n2. How Is Your Life Today? [05:30]\n3. Four Chords That Made A Million [08:15]\n4. Shemovedon [11:51]\n5. Last Chance To Evacuate Planet Earth Before It Is Recycled [17:07]\n6. The Rest Will Flow [21:56]\n7. Hatesong [25:10]\n8. Where We Would Be [33:34]\n9. Russia On Ice [37:49]\n10. Feel So Low [50:51]",
+    "thumbnails": {
+      "default": {
+        "url": "https://i.ytimg.com/vi/j3saYjV6UtA/default.jpg",
+        "width": 120,
+        "height": 90
+      },
+      "medium": {
+        "url": "https://i.ytimg.com/vi/j3saYjV6UtA/mqdefault.jpg",
+        "width": 320,
+        "height": 180
+      },
+      "high": {
+        "url": "https://i.ytimg.com/vi/j3saYjV6UtA/hqdefault.jpg",
+        "width": 480,
+        "height": 360
+      },
+      "standard": {
+        "url": "https://i.ytimg.com/vi/j3saYjV6UtA/sddefault.jpg",
+        "width": 640,
+        "height": 480
+      },
+      "maxres": {
+        "url": "https://i.ytimg.com/vi/j3saYjV6UtA/maxresdefault.jpg",
+        "width": 1280,
+        "height": 720
+      }
+    },
+    "channelTitle": "AlbumFullLyrics",
+    "tags": [
+      "Full",
+      "Album",
+      "Lyrics",
+      "Lightbulb Sun",
+      "How is your life today?",
+      "Four Chords That Made a Million",
+      "Shesmovedon",
+      "Last Chance to Evacuate Planet Earth Before It Is Recycled",
+      "The Rest Will Flow",
+      "Hatesong",
+      "Where We Would Be",
+      "Russia On Ice",
+      "Feel So Low",
+      "Prog",
+      "Progressive Rock",
+      "Music",
+      "Porcupine Tree"
+    ],
+    "categoryId": "22",
+    "liveBroadcastContent": "none",
+    "localized": {
+      "title": "Porcupine Tree - Lightbulb Sun (Full Album Lyrics)",
+      "description": "Here are the lyrics to the entire album of Lightbulb Sun by Porcupine Tree.\n\nLyrics taken from http://www.darklyrics.com/p/porcupinetree.html\n\nTracklist:\n1. Lightbulb Sun [00:00]\n2. How Is Your Life Today? [05:30]\n3. Four Chords That Made A Million [08:15]\n4. Shemovedon [11:51]\n5. Last Chance To Evacuate Planet Earth Before It Is Recycled [17:07]\n6. The Rest Will Flow [21:56]\n7. Hatesong [25:10]\n8. Where We Would Be [33:34]\n9. Russia On Ice [37:49]\n10. Feel So Low [50:51]"
+    }
+  },
+  "contentDetails": {
+    "duration": "PT56M11S",
+    "dimension": "2d",
+    "definition": "hd",
+    "caption": "false",
+    "licensedContent": false,
+    "projection": "rectangular"
+  },
+  "statistics": {
+    "viewCount": "11656",
+    "likeCount": "144",
+    "dislikeCount": "2",
+    "favoriteCount": "0",
+    "commentCount": "13"
+  }
+};
+
+export let VideoMockWithTracksLengthOnly = {
+  "kind": "youtube#video",
+  "etag": "\"m2yskBQFythfE4irbTIeOgYYfBU/D_NH1hiRA67N1elj7rpf6uJ3kjk\"",
+  "id": "4HqfLHTY4ws",
+  "snippet": {
+    "publishedAt": "2017-01-13T04:58:16.000Z",
+    "channelId": "UC8fqt_PDhDDszL5Zi8EauqA",
+    "title": "Ozric Tentacles - Waterfall Cities (full album)",
+    "description": "\"Coily\" (Ozric Tentacles) – 7:19\n    \"Xingu\" (Wynne, Seaweed, Geelani) – 7:27\n    \"Waterfall City\" (Wynne) – 11:03\n    \"Ch'ai?\" (Wynne, Seaweed, Geelani) – 5:03\n    \"Spiralmind\" (Wynne, Ozric Tentacles, Geelani) – 11:41\n    \"Sultana Detrii\" (Ozric Tentacles) – 9:18\n    \"Aura Borealis\" (Wynne) – 5:40\n\nBand personnel\n\n    Ed Wynne – guitar, synthesizers\n    Seaweed (Christopher Lenox-Smith) – synthesizers\n    John Egan – flute\n    Zia Geelani – bass guitar\n    Rad (Conrad Prince) – drums",
+    "thumbnails": {
+      "default": {
+        "url": "https://i.ytimg.com/vi/4HqfLHTY4ws/default.jpg",
+        "width": 120,
+        "height": 90
+      },
+      "medium": {
+        "url": "https://i.ytimg.com/vi/4HqfLHTY4ws/mqdefault.jpg",
+        "width": 320,
+        "height": 180
+      },
+      "high": {
+        "url": "https://i.ytimg.com/vi/4HqfLHTY4ws/hqdefault.jpg",
+        "width": 480,
+        "height": 360
+      },
+      "standard": {
+        "url": "https://i.ytimg.com/vi/4HqfLHTY4ws/sddefault.jpg",
+        "width": 640,
+        "height": 480
+      },
+      "maxres": {
+        "url": "https://i.ytimg.com/vi/4HqfLHTY4ws/maxresdefault.jpg",
+        "width": 1280,
+        "height": 720
+      }
+    },
+    "channelTitle": "Terminal Passage",
+    "tags": [
+      "Ozric Tentacles",
+      "Waterfall Cities",
+      "Full Album",
+      "Rock",
+      "Psychedelic Rock",
+      "Guitar",
+      "Synthesizer",
+      "Studio Album"
+    ],
+    "categoryId": "22",
+    "liveBroadcastContent": "none",
+    "localized": {
+      "title": "Ozric Tentacles - Waterfall Cities (full album)",
+      "description": "\"Coily\" (Ozric Tentacles) – 7:19\n    \"Xingu\" (Wynne, Seaweed, Geelani) – 7:27\n    \"Waterfall City\" (Wynne) – 11:03\n    \"Ch'ai?\" (Wynne, Seaweed, Geelani) – 5:03\n    \"Spiralmind\" (Wynne, Ozric Tentacles, Geelani) – 11:41\n    \"Sultana Detrii\" (Ozric Tentacles) – 9:18\n    \"Aura Borealis\" (Wynne) – 5:40\n\nBand personnel\n\n    Ed Wynne – guitar, synthesizers\n    Seaweed (Christopher Lenox-Smith) – synthesizers\n    John Egan – flute\n    Zia Geelani – bass guitar\n    Rad (Conrad Prince) – drums"
+    }
+  },
+  "contentDetails": {
+    "duration": "PT57M21S",
+    "dimension": "2d",
+    "definition": "hd",
+    "caption": "false",
+    "licensedContent": false,
+    "projection": "rectangular"
+  },
+  "statistics": {
+    "viewCount": "8174",
+    "likeCount": "78",
+    "dislikeCount": "4",
+    "favoriteCount": "0",
+    "commentCount": "10"
+  }
+};